[cypress] Breakup and remove commands.js
Breaks the file into multiple files of two different categories:

- commands
- assertions
