+++ cypress/support/index.js
@@ -11,10 +11,41 @@
 //
 // You can read more here:
 // https://on.cypress.io/configuration
+
+
+// ***********************************************
+// Below shows you how to create various custom
+// commands and overwrite existing commands.
+//
+// For more comprehensive examples of custom
+// commands please read more here:
+// https://on.cypress.io/custom-commands
+// ***********************************************
+//
+//
+// -- This is a parent command --
+// Cypress.Commands.add("login", (email, password) => { ... })
+//
+//
+// -- This is a child command --
+// Cypress.Commands.add("drag", { prevSubject: 'element'}, (subject, options) => { ... })
+//
+//
+// -- This is a dual command --
+// Cypress.Commands.add("dismiss", { prevSubject: 'optional'}, (subject, options) => { ... })
+//
+//
+// -- This will overwrite an existing command --
+// Cypress.Commands.overwrite("visit", (originalFn, url, options) => { ... })
 // ***********************************************************
 
-// Import commands.js using ES2015 syntax:
-import './commands'
+// Commands
+import './commands/gtl.js'
+import './commands/login.js'
+import './commands/menu.js'
+import './commands/search.js'
+import './commands/toolbar.js'
 
-// Alternatively you can use CommonJS syntax:
-// require('./commands')
+// Assertions
+import './assertions/expect_explorer_title.js'
+import './assertions/expect_show_list_title.js'
